% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOI_time.R
\name{AOI_time}
\alias{AOI_time}
\title{Analysis of time spent in areas of interest}
\usage{
AOI_time(data, data_type = NULL, AOIs, as_prop = FALSE, trial_time = NULL)
}
\arguments{
\item{data}{A dataframe  of either fixation data (from fix_dispersion) or raw data}

\item{data_type}{Whether data is a fixation ("fix") or raw data ("raw")}

\item{AOIs}{A dataframe of areas of interest (AOIs), with one row per AOI (name, x, y, width_radius, height).}

\item{as_prop}{whether to return time in AOI as a proportion of the total time of trial}

\item{trial_time}{needed if as_prop is set to TRUE. a vector of the time taken in each trial. Equal to the length of x trials by y participants in the dataset}
}
\value{
a dataframe containing the time on the passed AOIs for each trial. One column for each AOI separated by trial.
}
\description{
Analyses total time on defined AOI regions across trials. Works with fixation and raw data as the input (must use one or the other, not both).
}
\details{
Analyses data separately for each unique combination of values in \code{pID} and \code{trial}. Returned values can be absolute time or proportion of time over the period.
}
\examples{

\donttest{
data <- combine_eyes(HCL)
fix_d <- fixation_dispersion(data)

# fixation data
AOI_time(data = fix_d, data_type = "fix", AOIs = HCL_AOIs)

#raw data
AOI_time(data = data, data_type = "raw", AOIs = HCL_AOIs)

#as proportional data
AOI_time(data = fix_d, data_type = "fix", AOIs = HCL_AOIs,
         as_prop = TRUE, trial_time = HCL_behavioural$RT)
}
}
