% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOI_time_binned.R
\name{AOI_time_binned}
\alias{AOI_time_binned}
\title{Binned time analysis of area of interest entries}
\usage{
AOI_time_binned(
  data,
  AOIs,
  bin_length = NULL,
  max_time = NULL,
  as_prop = FALSE
)
}
\arguments{
\item{data}{A dataframe of raw data}

\item{AOIs}{A dataframe of areas of interest (AOIs), with one row per AOI (name, x, y, width_radius, height).}

\item{max_time}{maximum length of time to use, default is total trial length}

\item{as_prop}{whether to return time in AOI as a proportion of the total time of trial}
}
\value{
a dataframe containing the time on the passed AOIs for each trial. One column for each AOI separated by trial.
}
\description{
Analyses total time on defined AOI regions across trials separated into bins. Works with raw data as the input.
Data can be separated into bins of a given length of time and the number of bins per trial is calculated automatically, keeping the bin length
consistent across varying lengths of trial. Any data that cannot fill a bin (typically the last few milliseconds of the trial) are dropped to
ensure that bins are of a consistent length
}
\details{
AOI_time_binned can take either single participant data or multiple participants, where participants are demarcated by values in the "pID" column.
}
\examples{

\donttest{
data <- combine_eyes(HCL)


#with bins of 100ms each and only for the first 2000ms
AOI_time_binned(data = data, AOIs = HCL_AOIs, 
    bin_length = 100, max_time = 2000)
}

}
