% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AOIs_WM}
\alias{AOIs_WM}
\title{Example AOIs for use with example_raw_WM}
\format{
A data frame with 4 rows and 4 variables:
\describe{
\item{x}{centred x coordinate of the AOI}
\item{y}{centred y coordinate of the AOI}
\item{width_radius}{either the width of the AOI, or the radius for circular AOIs}
\item{height}{the height of the AOI; should be NA for circular AOIs}
...
}
}
\usage{
AOIs_WM
}
\description{
This dataframe contains 4 rectangular areas of interest (AOIs), set out for use with the example_raw_WM dataset. Values are in pixels.
}
\keyword{datasets}
