% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_counterbalance}
\alias{example_counterbalance}
\title{Example dataset from a dataset that includes counterbalancing}
\format{
A data frame with 80,577 rows and 10 variables:
\describe{
\item{time}{timestamp of the sample}
\item{x}{x coordinate of the eye}
\item{y}{y coordinate of the eye}
\item{trial}{trial number}
\item{P_cue}{Are these necessary columns?}
\item{NP_cue}{Are these necessary columns?}
\item{cue_order}{whether the predictive cue os presented on the left (1) or the right (2)}
\item{correct_out}{Are these necessary columns?}
\item{accuracy}{accuracy of selecting the correct cue}
\item{RT}{response time}
...
}
}
\usage{
example_counterbalance
}
\description{
This dataframe contains 96 trials where the predictive cue is counterbalanced and presented either on the left or the right.
}
\keyword{datasets}
