% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_raw_fix}
\alias{example_raw_fix}
\title{Example eye data from a simple contingency learning task}
\format{
A data frame with 77,669 rows and 4 variables:
\describe{
\item{time}{timestamp of the sample}
\item{x}{x coordinate of the eye}
\item{y}{y coordinate of the eye}
\item{trial}{trial number}
...
}
}
\usage{
example_raw_fix
}
\description{
A dataset containing raw data samples for 72 trials of a simple learning experiment,
with several clearly defined AOIs on the screen, producing many distinct fixations each trial.
Data were collected from a 300 Hz Tobii eye-tracker.
This is an example of the format raw data needs to be in in order to work with the functions
in eyetools.
}
\keyword{datasets}
