% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_raw_psy}
\alias{example_raw_psy}
\title{Example eye data}
\format{
A data frame with 77,669 rows and 4 variables:
\describe{
  \item{time}{timestamp of the sample}
  \item{x}{x coordinate of the eye}
  \item{y}{y coordinate of the eye}
  \item{trial}{trial number}
  ...
}
}
\usage{
example_raw_psy
}
\description{
A dataset containing raw data samples for 72 trials of a simple learning experiment,
which several clearly defined AOIs producing several distinct fixations each trial.
This is the format raw data needs to be in in order to work with the functions
within the "Processing" step of the workflow.
}
\keyword{datasets}
