% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_raw_WM}
\alias{example_raw_WM}
\title{Example eye data from working memory task}
\format{
A data frame with 24205 rows and 4 variables:
\describe{
\item{time}{timestamp of the sample}
\item{x}{x coordinate of the eye}
\item{y}{y coordinate of the eye}
\item{trial}{trial number}
...
}
}
\usage{
example_raw_WM
}
\description{
A dataset containing raw data samples for 100 trials of a working memory task. The screen contained
four objects (coloured shapes), positioned at the corners of a square and participants spent 2 seconds encoding the stimuli.
Data were collected from a 120 Hz Tobii eye-tracker.
This is an example of the format raw data needs to be in in order to work with the functions
in eyetools.
}
\keyword{datasets}
