% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_two_eyes_raw}
\alias{example_two_eyes_raw}
\title{Example eye data}
\format{
A data frame with X rows and 4 variables:
\describe{
  \item{time}{timestamp of the sample}
  \item{x_left}{x coordinate of the left eye}
  \item{y_left}{y coordinate of the left eye}
  \item{val_left}{validity code of the left eye}
  \item{x_right}{x coordinate of the right eye}
  \item{y_right}{y coordinate of the right eye}
  \item{val_right}{validity code of the right eye}
  \item{trial}{trial number}
  ...
}
}
\usage{
example_two_eyes_raw
}
\description{
A dataset containing raw data samples for 10 trials. The data are from two eyes,
and therefore this dataset is to be used as an example for the "combine_eyes" function.
}
\keyword{datasets}
