% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_dispersion_tolerance.R
\name{explore_dispersion_tolerance}
\alias{explore_dispersion_tolerance}
\title{Fixation detection using inverse saccade identification}
\usage{
explore_dispersion_tolerance(
  data,
  algorithm = c("dispersion", "VTI", "both"),
  tolerance_min = 75,
  tolerance_max = 120,
  step = 5
)
}
\arguments{
\item{data}{A dataframe with raw data (time, x, y, trial) for one participant}

\item{algorithm}{A string of either "dispersion", "VTI", or "both" to specify which algorithm to test}

\item{tolerance_min}{the minimum dispersion tolerance to test}

\item{tolerance_max}{the maximum dispersion tolerance to test}

\item{step}{the step to take between the minimum and maximum tolerance values to explore}
}
\description{
Description
}
