% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyetools-package.R
\docType{package}
\name{eyetools-package}
\alias{eyetools}
\alias{eyetools-package}
\title{eyetools: Analyse Eye Data}
\description{
Enables the automation of actions across the pipeline, including initial steps of transforming binocular data and gap repair to event-based processing such as fixations, saccades, and entry/duration in Areas of Interest (AOIs). It also offers visualisation of eye movement and AOI entries. These tools take relatively raw (trial, time, x, and y form) data and can be used to return fixations, saccades, and AOI entries and time spent in AOIs. As the tools rely on this basic data format, the functions can work with data from any eye tracking device. Implements fixation and saccade detection using methods proposed by Salvucci and Goldberg (2000) \doi{10.1145/355017.355028}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tombeesley.github.io/eyetools/}
  \item Report bugs at \url{https://github.com/tombeesley/eyetools/issues}
}

}
\author{
\strong{Maintainer}: Tom Beesley \email{t.beesley@lancaster.ac.uk}

Authors:
\itemize{
  \item Tom Beesley \email{t.beesley@lancaster.ac.uk}
  \item Matthew Ivory \email{matthew.ivory@lancaster.ac.uk}
}

}
\keyword{internal}
