% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dispersion.R
\name{fix_dispersion}
\alias{fix_dispersion}
\title{Fixation detection by dispersion method}
\usage{
fix_dispersion(data, min_dur = 50, disp_tol = 30, report = FALSE)
}
\arguments{
\item{data}{dataframe with columns time, x, y, trial (the standardised raw data form for eyeproc)}
}
\value{

}
\description{
Detects fixations according to the method proposed by S & G
}
\examples{

}
