% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dispersion_old.R
\name{fix_dispersion_old}
\alias{fix_dispersion_old}
\title{Fixation detection by dispersion method OLD - no longer in us}
\usage{
fix_dispersion_old(data, min_dur = 150, disp_tol = 100)
}
\arguments{
\item{data}{dataframe with columns time, x, y, trial (the standardised raw data form for eyeproc)}
}
\value{

}
\description{
Detects fixations according to the method proposed by S & G
}
\examples{

}
\keyword{internal}
