% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_dispersion_SLOW.R
\name{fix_dispersion_SLOW}
\alias{fix_dispersion_SLOW}
\title{Fixation detection by dispersion method}
\usage{
fix_dispersion_SLOW(data, min_dur = 150, disp_tol = 150)
}
\arguments{
\item{data}{dataframe with columns time, x, y, trial (the standardised raw data form for eyeproc)}
}
\value{

}
\description{
Detects fixations according to the method proposed by S & G
}
\examples{

}
\keyword{internal}
