% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HCL_AOIs}
\alias{HCL_AOIs}
\title{Example AOIs for use with HCL}
\format{
A data frame with 3 rows and 5 variables:
\describe{
\item{name}{custom name for the AOI}
\item{x}{centred x coordinate of the AOI}
\item{y}{centred y coordinate of the AOI}
\item{width_radius}{either the width of the AOI, or the radius for circular AOIs}
\item{height}{the height of the AOI; should be NA for circular AOIs}
...
}
}
\usage{
HCL_AOIs
}
\description{
This dataframe contains three rectangular areas of interest (AOIs), set out for use with the HCL dataset.
Values are in pixels. This is the standard AOI format required by many functions in the package. A template dataframe
can be created using create_AOI_df()
}
\keyword{datasets}
