% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HCL_behavioural}
\alias{HCL_behavioural}
\title{Example dataset of behavioural data to compliment dataset HCL. This contains information on stimuli
(such as the side the predictive cue was presented on) as well as response data, including accuracy and response times}
\format{
A dataframe of 12 observations and eight variables
\describe{
\item{pNum}{participant number}
\item{trial}{trial number}
\item{P_cue}{Are these necessary columns?}
\item{NP_cue}{Are these necessary columns?}
\item{cue_order}{whether the predictive cue os presented on the left (1) or the right (2)}
\item{correct_out}{NAre these necessary columns?}
\item{accuracy}{response accuracy}
\item{RT}{response time in milliseconds}
...
}
}
\usage{
HCL_behavioural
}
\description{
Example dataset of behavioural data to compliment dataset HCL. This contains information on stimuli
(such as the side the predictive cue was presented on) as well as response data, including accuracy and response times
}
\keyword{datasets}
