% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_get_event.R
\name{hdf5_get_event}
\alias{hdf5_get_event}
\title{Get messages stored in TOBII-generated HDF5 files}
\usage{
hdf5_get_event(filename)
}
\arguments{
\item{filename}{the hdf5 file generated from TOBII}
}
\value{
A dataframe of message events as recorded by TOBII eye trackers
}
\description{
A function to get the message event files from a TOBII-generated hdf5 file to dataframe.
Used when a Psychopy experiment uses the io.sendMessageEvent() to record events
}
\examples{
\dontrun{
raw_data <- hdf5_get_event("example_TOBII.hdf5")
}
}
