% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_to_csv.R
\name{hdf5_to_csv}
\alias{hdf5_to_csv}
\title{hdf5_to_csv}
\usage{
hdf5_to_csv(filename)
}
\arguments{
\item{filename}{the hdf5 file generated from TOBII}
}
\value{
A list of csv files collected from the eyetracker content, if only one eyetracking event is present, return this as a csv file
}
\description{
A function to convert TOBII-generated hdf5 files to csv
}
\examples{
\dontrun{
raw_data <- hdf5_to_csv("example_TOBII.hdf5")
}
}
