% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdf5_to_df.R
\name{hdf5_to_df}
\alias{hdf5_to_df}
\title{Convert TOBII-generated HDF5 files to dataframe}
\usage{
hdf5_to_df(filename)
}
\arguments{
\item{filename}{the hdf5 file generated from TOBII}
}
\value{
A list of dataframes collected from the eyetracker content, if only one eyetracking event is present, return this as a single dataframe
}
\description{
A function to convert TOBII-generated hdf5 files to a dataframe
}
\examples{
\dontrun{
raw_data <- hdf5_to_df("example_TOBII.hdf5")
}
}
