% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_AOI_growth.R
\name{plot_AOI_growth}
\alias{plot_AOI_growth}
\title{Plots absolute or proportional time spent in AOIs over time}
\usage{
plot_AOI_growth(
  data = NULL,
  pID_values = NULL,
  trial_values = NULL,
  AOIs = NULL,
  type = "abs",
  plot_time_not_in_AOI = FALSE
)
}
\arguments{
\item{data}{raw data in standard raw data form (time, x, y, trial)}

\item{pID_values}{specify particular values within 'pID' to plot data from certain participants}

\item{trial_values}{can be used to select particular trials within the data}

\item{AOIs}{A dataframe of areas of interest (AOIs), with one row per AOI (name, x, y, width_radius, height)}

\item{type}{either "abs" (absolute) or "prop" (proportion)}

\item{plot_time_not_in_AOI}{boolean as to whether to include proportion of time spent outside AOIs}
}
\value{
a plot of the raw data
}
\description{
A visualisation tool for plotting the changes in defined AOI regions across a single trial time.
}
\examples{

\donttest{
data <- combine_eyes(HCL)
data <- interpolate(data)
# plot absolute and then proportional
plot_AOI_growth(data = data, AOIs = HCL_AOIs, type = "abs")
plot_AOI_growth(data = data, AOIs = HCL_AOIs, type = "prop")
}

}
