% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_seq.R
\name{plot_seq}
\alias{plot_seq}
\title{Plot of raw data over time}
\usage{
plot_seq(
  data = NULL,
  AOIs = NULL,
  trial_values = NULL,
  pID_values = NULL,
  bg_image = NULL,
  flip_y = FALSE,
  plot_header = FALSE,
  bin_time = NULL,
  bin_range = NULL
)
}
\arguments{
\item{data}{A dataframe with raw data. If multiple trials are used, then one trial is sampled at random.}

\item{AOIs}{A dataframe of areas of interest (AOIs), with one row per AOI (name, x, y, width_radius, height).}

\item{trial_values}{specify particular values within 'trial' to plot data from certain trials}

\item{pID_values}{specify particular values within 'pID' to plot data from certain participants}

\item{bg_image}{The filepath of a PNG image to be added to the plot, for example to show a screenshot of the task.}

\item{flip_y}{reverse the y axis coordinates (useful if origin is top of the screen)}

\item{plot_header}{display the header title text which explains graphical features of the plot.}

\item{bin_time}{if wanting to split data into bins, the time (in ms) for each bin of data to be displayed}

\item{bin_range}{if wanting to split data into bins, the first and last bin to be display, e.g., c(1,5)}
}
\value{
a plot of the raw data representing changes over time
}
\description{
A tool for visualising the timecourse of raw data over a single trial. Participant and trial values can be selected from the data.
If values for these parameters are not provided then a single participant and a single trial will be sampled at random.
Data can be split into bins by time or by the number of bins.
}
\examples{
data <- combine_eyes(HCL)

# plot the raw data
plot_seq(data)

# with AOIs
plot_seq(data, AOIs = HCL_AOIs)

# plot raw data with bins
plot_seq(data, bin_time = 500)

}
