% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{smoother}
\alias{smoother}
\title{Smoothing of raw data}
\usage{
smoother(data, span = 0.05, plot = FALSE)
}
\arguments{
\item{data}{A dataframe with raw data (time, x, y, trial) for one participant}

\item{span}{From stats::loess. The parameter alpha which controls the degree of smoothing.}

\item{plot}{whether to plot the raw and smoothed plot for inspection}
}
\value{
a dataframe of the same shape as the input data
}
\description{
A wrapper for the stats::loess function, with default parameters suitable for smoothing raw eye data
}
\details{
It can take either single participant data or multiple participants where there is a variable for unique participant identification.
The function looks for an identifier named \code{participant_col} by default and will treat this as multiple-participant data as default,
if not it is handled as single participant data, or the participant_col needs to be specified
}
\examples{
data <- combine_eyes(HCL)

smoother(data)

#with an inspection plot
smoother(data, span = .02, plot = TRUE)

}
