% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_plot.R
\name{spatial_plot}
\alias{spatial_plot}
\title{Plot raw data and fixations}
\usage{
spatial_plot(
  raw_data = NULL,
  fix_data = NULL,
  AOIs = NULL,
  res = NULL,
  flip_y = FALSE,
  show_fix_order = TRUE
)
}
\arguments{
\item{raw_data}{data in standard raw data form (time, x, y, trial)}

\item{fix_data}{data output from fixation function (start, end, x, y, duration, trial)}

\item{AOIs}{A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width/diameter, height/NA).}

\item{res}{resolution of the display to be shown, as a vector (xmin, xmax, ymin, ymax)}

\item{flip_y}{reverse the y axis coordinates (useful if origin is top of the screen)}

\item{show_fix_order}{label the fixations in the order they were made}
}
\value{

}
\description{
A tool for visualising raw eye-data and processed fixations. Can pass both raw data and fixation data. Fixations can be labeled
in the order they were made. Can overlay areas of interest (AOIs) and customise the resolution.
}
\examples{

}
