% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_properties_get.R
\name{tracker_properties_get}
\alias{tracker_properties_get}
\title{Get Current Eyetracker Properties}
\usage{
tracker_properties_get()
}
\value{
A named list of the current tracker parameters.
}
\description{
Returns a list containing the current global eyetracker settings.
}
