% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracker_properties_set.R
\name{tracker_properties_set}
\alias{tracker_properties_set}
\title{Set the properties of the eye-tracker that was used in recording the data}
\usage{
tracker_properties_set(
  sample_frequency = NULL,
  viewing_distance_cm = NULL,
  screen_width_cm = NULL,
  screen_height_cm = NULL,
  screen_width_pixels = NULL,
  screen_height_pixels = NULL
)
}
\arguments{
\item{sample_frequency}{Numeric frequency.}

\item{viewing_distance_cm}{Distance in cm.}

\item{screen_width_cm}{Width in cm.}

\item{screen_height_cm}{Height in cm.}

\item{screen_width_pixels}{Width in pixels.}

\item{screen_height_pixels}{Height in pixels.}
}
\description{
This function is used to specify the properties of the eye-tracker used to record the data.
Users can update one or more parameters. Unnamed parameters retain previous values.
If these are not set by the user, then default parameters are set and a warning is issued in every call to relevant functions
}
