% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOI_seq.R
\name{AOI_seq}
\alias{AOI_seq}
\title{Sequence analysis of area of interest entries}
\usage{
AOI_seq(
  fix_data = NULL,
  raw_data = NULL,
  AOIs,
  AOI_names = NULL,
  sample_rate = NULL
)
}
\arguments{
\item{fix_data}{A dataframe with fixation data (from fix_dispersion)}

\item{raw_data}{A dataframe with raw data}

\item{AOIs}{A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width_radius, height).}

\item{AOI_names}{An optional vector of AOI names to replace the default "AOI_1", "AOI_2", etc.}

\item{sample_rate}{Optional sample rate of the eye-tracker (Hz) for use with raw_data. If not supplied, the sample rate will be estimated from the time column and the number of samples.}
}
\value{
a dataframe containing the sequence of entries into AOIs on each trial
}
\description{
Analyses the sequence of entries into defined AOI regions across trials. Works with fixation and raw data as the input (must use one or the other, not both).
}
\examples{
fix_d <- fix_dispersion(eyetools::example_raw_WM)
AOI_seq(fix_d, eyetools::AOIs_WM)

}
