% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOI_time.R
\name{AOI_time}
\alias{AOI_time}
\title{Area of interest analysis at the trial level}
\usage{
AOI_time(data, AOIs)
}
\arguments{
\item{data}{A dataframe with fixation data (from fix_dispersion)}
}
\value{

}
\description{
Analyses time on AOI regions
}
\examples{

}
