% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOI_time.R
\name{AOI_time}
\alias{AOI_time}
\title{Area of interest analysis at the trial level}
\usage{
AOI_time(fix_data = NULL, raw_data = NULL, AOIs, AOI_names = NULL)
}
\arguments{
\item{AOIs}{A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width_radius, height).}

\item{AOI_names}{An optional vector of AOI names to replace the default "AOI_1", "AOI_2", etc.}

\item{data}{A dataframe with fixation data (from fix_dispersion)}
}
\value{
a dataframe containing the time on the passed AOIs for each trial
}
\description{
Analyses total time on defined AOI regions across trials. Currently only works with fixation data as the input.
}
\examples{
fix_d <- fix_dispersion(eyetools::example_raw_WM)
AOI_time(fix_d, eyetools::AOIs_WM)

}
