% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOI_time.R
\name{AOI_time}
\alias{AOI_time}
\title{Time analysis of area of interest entries}
\usage{
AOI_time(
  fix_data = NULL,
  raw_data = NULL,
  AOIs,
  AOI_names = NULL,
  sample_rate = NULL
)
}
\arguments{
\item{fix_data}{A dataframe with fixation data (from fix_dispersion)}

\item{raw_data}{A dataframe with raw data}

\item{AOIs}{A dataframe of areas of interest (AOIs), with one row per AOI (x, y, width_radius, height).}

\item{AOI_names}{An optional vector of AOI names to replace the default "AOI_1", "AOI_2", etc.}

\item{sample_rate}{Optional sample rate of the eye-tracker (Hz) for use with raw_data. If not supplied, the sample rate will be estimated from the time column and the number of samples.}
}
\value{
a dataframe containing the time on the passed AOIs for each trial
}
\description{
Analyses total time on defined AOI regions across trials. Works with fixation and raw data as the input (must use one or the other, not both).
}
\examples{
fix_d <- fixation_dispersion(example_raw_WM)
AOI_time(fix_data = fix_d, AOIs = AOIs_WM)
AOI_time(raw_data = example_raw_WM, AOIs = AOIs_WM, sample_rate = 120)

}
