% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_transform.R
\name{conditional_transform}
\alias{conditional_transform}
\title{conditional_transform}
\usage{
conditional_transform(
  data,
  flip = c("x", "y"),
  cond_column,
  cond_values,
  resolution_x = 1920,
  resolution_y = 1080
)
}
\arguments{
\item{data}{a dataframe that includes columns of time, x, y, trial, and cue_order. Cue order should currently be in the state of 1,2 which refer to the counterbalancing of trials.}

\item{flip}{either x or y to specify a flip across either the vertical axis or horizontal axis}

\item{cond_column}{a column name, on which the flips are conditional}

\item{cond_values}{which values in this column result in the flip}

\item{resolution_x}{screen size in pixels for the x axis}

\item{resolution_y}{screen size in pixels for the y axis}
}
\value{
a dataframe of the equivalent format as the input data
}
\description{
A function to perform conditional transformations of the x/y raw data.
The function takes the dataframe and performs a single axis flip based on the values specificed in the cond_column.
The primary use of this function is to correct or normalise the data when counterbalancing stimulus placement within experiments (e.g., having a target stimulus appear on the left and right equally often)
}
\examples{
conditional_transform(example_counterbalance, flip = "x", cond_column = "cue_order", cond_values = 2)

}
