% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conditional_transform.R
\name{conditional_transform}
\alias{conditional_transform}
\title{conditional_transform}
\usage{
conditional_transform(
  data,
  flip = c("x", "y"),
  resolution_x = 1920,
  resolution_y = 1080
)
}
\arguments{
\item{data}{a dataframe that includes columns of time, x, y, trial, and cue_order. Cue order should currently be in the state of 1,2 which refer to the counterbalancing of trials.}

\item{flip}{either x or y to specify a flip across either the vertical axis or horizontal axis}

\item{resolution_x}{screen size in pixels for the x axis}

\item{resolution_y}{screen size in pixels for the y axis}
}
\description{
A function to counterbalance trials. The function takes the dataframe and performs a single axis flip to correct counterbalancing.
}
\examples{
\dontrun {
example_raw_HCL <- example_raw_HCL |> mutate(x = ifelse(x > 1920, NA, x),
y = ifelse(y > 1080, NA, y))
data <- example_raw_HCL |> left_join(example_resp_HCL, by = join_by(trial))
conditional_transform(data, flip = "x")
}


}
