% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_AOI_df.R
\name{create_AOI_df}
\alias{create_AOI_df}
\title{Create a blank data frame for populating with AOIs}
\usage{
create_AOI_df(num_AOIs = 3, AOI_data = NULL, AOI_names = NULL)
}
\arguments{
\item{num_AOIs}{number of AOIs, setting the number of rows}

\item{AOI_data}{a list containing data for each AOI, ordered by x, y, width_radius, and height (NA if circular)}

\item{AOI_names}{a vector of names for the AOIs specified in AOI_data}
}
\value{
an AOI dataframe in the format required for several AOI functions in eyetools
}
\description{
Used to define a set of AOIs for use across various functions within the eyetools package.
}
\examples{
# create an empty data frame with 3 AOIs
create_AOI_df(3)

# create an AOI dataframe with data, the second of which is circular, with names
create_AOI_df(num_AOIs = 3, 
              AOI_data = list(c(460,840,400,300), c(1460,840,400,NA), c(960,270,300,500)),
              AOI_names = c("AOI_1", "AOI_2", "AOI_3"))
}
