% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_AOI_df.R
\name{create_AOI_df}
\alias{create_AOI_df}
\title{Create a blank data frame for populating with AOIs}
\usage{
create_AOI_df(num_AOIs = 3)
}
\arguments{
\item{num_AOIs}{number of AOIs, setting the number of rows}
}
\value{
a dataframe in the standard format required for eyetools
}
\description{
Create a blank data frame for populating with AOIs
}
\examples{
# create a data frame with 4 AOIs
create_AOI_df(4)

}
