% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyetools-package.R
\docType{package}
\name{eyetools-package}
\alias{eyetools}
\alias{eyetools-package}
\title{eyetools: Tools for Eye Data Analysis}
\description{
A set of tools for easier eye data processing. It enables the automation of core actions across the pipeline from basic data formats of trial, time, x, and y coordinates, and standardises the processing of binocular data, fixing missing data, smoothing data, and plotting eye gaze information. The toolset also offers event-related processing, such as calculating fixations, saccades, and time in Areas of Interest. The tools are designed to work with data from any eye tracking device.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tombeesley.github.io/eyetools/}
}

}
\author{
\strong{Maintainer}: Beesley Tom \email{t.beesley@lancaster.ac.uk}

Authors:
\itemize{
  \item Ivory Matthew \email{matthew.ivory@lancaster.ac.uk}
}

}
\keyword{internal}
