% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyetools-package.R
\docType{package}
\name{eyetools-package}
\alias{eyetools}
\alias{eyetools-package}
\title{eyetools: Tools for eye data analysis}
\description{
A set of tools for eye data processing. The intended flow is from data processing, to fixation analysis, to data visualisation and trial level summaries.
}
\keyword{internal}
