% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyetools-package.R
\docType{package}
\name{eyetools-package}
\alias{eyetools}
\alias{eyetools-package}
\title{eyetools: Tools for Eye Data Analysis}
\description{
A set of tools for easier eye data processing. It enables automation of actions across the pipeline starting from combining binocular data, repairing missing data, through to event-related processing (fixations, saccades, time in AOIs), to data visualisation. These tools take relatively raw (trial, time, x, and y form) data and can be used to return fixations, saccades, and AOI entries and time spent in AOIs. As the tools rely on this basic data format, the functions can work with data from any eye tracking device.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tombeesley.github.io/eyetools/}
}

}
\author{
\strong{Maintainer}: Beesley Tom \email{matthew.ivory@lancaster.ac.uk}

Authors:
\itemize{
  \item Ivory Matthew \email{matthew.ivory@lancaster.ac.uk}
}

}
\keyword{internal}
