% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eyetools-package.R
\docType{package}
\name{eyetools-package}
\alias{eyetools}
\alias{eyetools-package}
\title{eyetools: Tools for Eye Data Analysis}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A set of tools for eye data processing. The intended flow is from data processing, to fixation analysis, to data visualisation and trial level summaries.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://tombeesley.github.io/eyetools/}
}

}
\author{
\strong{Maintainer}: Beesley Tom \email{t.beesley@lancaster.ac.uk}

Authors:
\itemize{
  \item Ivory Matthew \email{matthew.ivory@lancaster.ac.uk}
}

}
\keyword{internal}
