% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixation_dispersion.R
\name{fixation_dispersion}
\alias{fixation_dispersion}
\title{Fixation detection by dispersion method}
\usage{
fixation_dispersion(
  data,
  min_dur = 150,
  disp_tol = 100,
  run_interp = TRUE,
  NA_tol = 0.25,
  progress = TRUE
)
}
\arguments{
\item{data}{A dataframe with raw data (time, x, y, trial) for one participant (the standardised raw data form for eyeproc)}

\item{min_dur}{Minimum duration (in milliseconds) of period over which fixations are assessed}

\item{disp_tol}{Maximum tolerance (in pixels) for the dispersion of values allowed over fixation period}

\item{run_interp}{include a call to eyetools::interpolate on each trial}

\item{NA_tol}{the proportion of NAs tolerated within any window of samples that is evaluated as a fixation}

\item{progress}{Display a progress bar}
}
\value{
a dataframe containing each detected fixation by trial, with mean x/y position in pixel, start and end times, and duration.
}
\description{
Detects fixations according to a method similar to that proposed by Salvucci & Goldberg (1996).
Evaluates the absolute range of values of both x and y coordinates. Looks for sufficient periods
in which the ranges are both below the specified dispersion tolerance. NAs are considered breaks
in the data and are not permitted within a valid fixation period.
}
\examples{
fixation_dispersion(example_raw_fix, disp_tol = 150)

}
