% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixation_VTI.R
\name{fixation_VTI}
\alias{fixation_VTI}
\title{Fixation detection using inverse saccade identification}
\usage{
fixation_VTI(
  data,
  sample_rate = NULL,
  threshold = 100,
  min_dur = 150,
  min_dur_sac = 20,
  disp_tol = 100,
  run_interp = TRUE,
  smooth = FALSE,
  progress = TRUE
)
}
\arguments{
\item{data}{A dataframe with raw data (time, x, y, trial) for one participant}

\item{sample_rate}{sample rate of the eye-tracker. If default of NULL, then it will be computed from the timestamp data and the number of samples}

\item{threshold}{velocity threshold (degrees of VA / sec) to be used for identifying saccades.}

\item{min_dur}{Minimum duration (in milliseconds) of period over which fixations are assessed}

\item{min_dur_sac}{Minimum duration (in milliseconds) for saccades to be determined}

\item{disp_tol}{Maximum tolerance (in pixels) for the dispersion of values allowed over fixation period}

\item{run_interp}{include a call to eyetools::interpolate on each trial.}

\item{smooth}{include a call to eyetools::smoother on each trial}

\item{progress}{Display a progress bar}
}
\value{
A list of the summary dataframe, the correlations conducted, and the data used for plotting the fixation bars
}
\description{
Determine fixations through the removal of saccades using the velocity threshold algorithm from Salvucci & Goldberg (1996).
Applies the algorithm used in VTI_saccade and removes the identified saccades before assessing whether separated fixations are outside of the dispersion tolerance.
If they are outside this tolerance, the fixation is treated as a new fixation regardless of the length of saccade separating them.
Compared to fixation_dispersion(), fixation_VTI() is more conservative in determining a fixation as smaller saccades are discounted and following data treated as a continued fixation (assuming within pixel tolerance using disp_tol).
Returns a summary of the saccades found per trial, including start and end coordinates, timing, duration, mean velocity, and peak velocity.
}
\examples{
# single trials:
fixation_VTI(example_raw_WM[example_raw_WM$trial == 15,])

# multiple trials:
data <- rbind(example_raw_WM[example_raw_WM$trial \%in\% c(3:10),])

fixation_VTI(data)
}
