% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate.R
\name{interpolate}
\alias{interpolate}
\title{Interpolation of missing data (NAs)}
\usage{
interpolate(
  data,
  vel_threshold = 35,
  maxgap = 150,
  method = "approx",
  report = FALSE,
  progress = TRUE
)
}
\arguments{
\item{data}{A dataframe with raw data (pID, time, x, y, trial), the standardised raw data form for eyetools}

\item{vel_threshold}{the maximum velocity that is tolerated between start and end points of periods of missing data}

\item{maxgap}{maximum time gap of consecutive trackloss to fill (in ms). Any longer gaps will be left unchanged (see zoo package)}

\item{method}{"approx" for linear interpolation or "spline" for cubic spline interpolation}

\item{report}{default is FALSE. If TRUE, then the return value is a list containing the returned data frame and the report.}

\item{progress}{Display a progress bar}
}
\value{
a dataframe of the same shape of the input data
}
\description{
Extends the zoo::na.approx and zoo::na.spline functions to include a report which provides
the proportion of missing data before and after the interpolation process. This is handy
for evaluating the effectiveness of the repair.
}
\details{
Analyses data separately for each unique combination of values in \code{pID} and \code{trial}.
}
\examples{
data <- combine_eyes(HCL)
interpolate(data, maxgap = 150)

}
