% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{smoother}
\alias{smoother}
\title{Smoothing of raw data}
\usage{
smoother(data, span = 0.1)
}
\arguments{
\item{data}{A dataframe with raw data (time, x, y, trial) for one participant}
}
\value{

}
\description{
A wrapper for the stats::loess function, with default parameters suitable for smoothing raw eye data
}
\examples{

}
